apps/api/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
httpx>=0.24.0        # HTTP client with timeout support
tenacity>=8.2.0      # Retry decorator